requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.40
langgraph>=0.0.20
langchain-text-splitters>=0.0.1

# LLM Providers
langchain-ollama>=0.0.1      # Ollama (local)
langchain-openai>=0.0.5      # OpenAI
langchain-anthropic>=0.1.0   # Anthropic/Claude

# Vector database
langchain-chroma>=0.1.0
chromadb>=0.4.0

# Document processing
pypdf>=3.0.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# HTTP
requests>=2.28.0

# Visualization
matplotlib>=3.7.0
